# src/hotel-management-api-spec.yaml
openapi: 3.0.3
info:
  title: Hotel Management System API
  description: |
    RESTful API for Hotel Operations Digitalization - Room Booking & Reservation Module.
    This API enables hotel staff to manage room availability, reservations, guest check-in/check-out, billing, and payments.
  version: 1.0.0
  contact:
    name: Hotel Management System Team
    email: support@hotelmanagementsystem.com
servers:
  - url: https://api.hotelmanagementsystem.com/v1
    description: Production server
  - url: https://staging-api.hotelmanagementsystem.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /rooms/availability:
    get:
      tags:
        - Rooms
      summary: Get room availability
      description: Retrieve real-time room availability based on check-in and check-out dates
      operationId: getRoomAvailability
      parameters:
        - name: checkInDate
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2024-03-15'
          description: Check-in date (YYYY-MM-DD)
        - name: checkOutDate
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2024-03-18'
          description: Check-out date (YYYY-MM-DD)
        - name: roomType
          in: query
          required: false
          schema:
            type: string
            enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          description: Filter by room type
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [AVAILABLE, BOOKED, MAINTENANCE]
          description: Filter by room status
      responses:
        '200':
          description: Room availability retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoomAvailability'
                  totalAvailable:
                    type: integer
                    example: 15
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /rooms/{roomId}:
    get:
      tags:
        - Rooms
      summary: Get room details
      description: Retrieve detailed information about a specific room
      operationId: getRoomById
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            example: RM-101
          description: Unique room identifier
      responses:
        '200':
          description: Room details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /reservations:
    post:
      tags:
        - Reservations
      summary: Create a new reservation
      description: Create a room reservation with guest details and stay information
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Room not available for selected dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: ROOM_NOT_AVAILABLE
                message: Selected room is not available for the specified dates
                timestamp: '2024-01-15T10:30:00Z'

    get:
      tags:
        - Reservations
      summary: Search reservations
      description: Search reservations by confirmation number, guest name, or date range
      operationId: searchReservations
      parameters:
        - name: confirmationNumber
          in: query
          schema:
            type: string
            example: CNF-2024-001234
        - name: guestName
          in: query
          schema:
            type: string
            example: John Doe
        - name: checkInDate
          in: query
          schema:
            type: string
            format: date
            example: '2024-03-15'
        - name: checkOutDate
          in: query
          schema:
            type: string
            format: date
            example: '2024-03-18'
        - name: status
          in: query
          schema:
            type: string
            enum: [CONFIRMED, CHECKED_IN, CHECKED_OUT, CANCELLED]
      responses:
        '200':
          description: Reservations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reservation'
                  totalCount:
                    type: integer
                    example: 25
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /reservations/{confirmationNumber}:
    get:
      tags:
        - Reservations
      summary: Get reservation details
      description: Retrieve detailed information about a specific reservation
      operationId: getReservationByConfirmationNumber
      parameters:
        - name: confirmationNumber
          in: path
          required: true
          schema:
            type: string
            example: CNF-2024-001234
      responses:
        '200':
          description: Reservation details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    Room:
      type: object
      properties:
        roomId:
          type: string
          example: RM-101
        roomNumber:
          type: string
          example: '101'
        roomType:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DELUXE
        status:
          type: string
          enum: [AVAILABLE, BOOKED, OCCUPIED, MAINTENANCE]
          example: AVAILABLE
        pricePerNight:
          type: number
          format: double
          example: 5000.00
        floor:
          type: integer
          example: 1
        amenities:
          type: array
          items:
            type: string
          example: [AC, TV, WiFi, Minibar]
        maxOccupancy:
          type: integer
          example: 2
    RoomAvailability:
      type: object
      properties:
        roomId:
          type: string
          example: RM-101
        roomNumber:
          type: string
          example: '101'
        roomType:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DELUXE
        status:
          type: string
          enum: [AVAILABLE, BOOKED, MAINTENANCE]
          example: AVAILABLE
        pricePerNight:
          type: number
          format: double
          example: 5000.00
        availableFrom:
          type: string
          format: date
          example: '2024-03-15'
        availableUntil:
          type: string
          format: date
          example: '2024-03-20'
    Guest:
      type: object
      required:
        - name
        - contactNumber
        - email
        - idProofType
        - idProofNumber
      properties:
        guestId:
          type: string
          example: GST-001234
        name:
          type: string
          example: John Doe
        contactNumber:
          type: string
          example: '+91-9876543210'
        email:
          type: string
          format: email
          example: john.doe@email.com
        idProofType:
          type: string
          enum: [PASSPORT, DRIVERS_LICENSE, NATIONAL_ID, AADHAAR]
          example: PASSPORT
        idProofNumber:
          type: string
          example: P1234567
        address:
          type: string
          example: 123 Main Street, City, Country
    CreateReservationRequest:
      type: object
      required:
        - roomType
        - checkInDate
        - checkOutDate
        - guest
      properties:
        roomType:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DELUXE
        roomId:
          type: string
          description: Optional specific room ID, if not provided system will assign available room
          example: RM-101
        checkInDate:
          type: string
          format: date
          example: '2024-03-15'
        checkOutDate:
          type: string
          format: date
          example: '2024-03-18'
        guest:
          $ref: '#/components/schemas/Guest'
        numberOfGuests:
          type: integer
          example: 2
        specialRequests:
          type: string
          example: Late check-in expected, extra pillows needed
    Reservation:
      type: object
      properties:
        confirmationNumber:
          type: string
          example: CNF-2024-001234
        roomId:
          type: string
          example: RM-101
        roomNumber:
          type: string
          example: '101'
        roomType:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DELUXE
        checkInDate:
          type: string
          format: date
          example: '2024-03-15'
        checkOutDate:
          type: string
          format: date
          example: '2024-03-18'
        guest:
          $ref: '#/components/schemas/Guest'
        numberOfGuests:
          type: integer
          example: 2
        numberOfNights:
          type: integer
          example: 3
        status:
          type: string
          enum: [CONFIRMED, CHECKED_IN, CHECKED_OUT, CANCELLED]
          example: CONFIRMED
        specialRequests:
          type: string
          example: Late check-in expected
        createdAt:
          type: string
          format: date-time
          example: '2024-03-10T14:30:00Z'
        checkInTime:
          type: string
          format: date-time
          nullable: true
          example: '2024-03-15T14:30:00Z'
        checkOutTime:
          type: string
          format: date-time
          nullable: true
          example: '2024-03-18T11:45:00Z'
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request parameters
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        details:
          type: array
          items:
            type: string
          example: ["checkInDate must be in the future", "checkOutDate must be after checkInDate"]
  responses:
    BadRequestError:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: BAD_REQUEST
            message: Invalid request parameters
            timestamp: '2024-01-15T10:30:00Z'
            details: ["checkInDate is required"]
    UnauthorizedError:
      description: Unauthorized - Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Invalid or expired authentication token
            timestamp: '2024-01-15T10:30:00Z'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Reservation not found
            timestamp: '2024-01-15T10:30:00Z'
...